# schemas/openapi.yaml
openapi: 3.1.0
info:
  title: HITL Protocol API
  version: "0.5"
  description: |
    The Open Standard for Human Decisions in Agent Workflows.

    HITL Protocol standardizes three-party flows between **Agent**, **Service**, and **Human**.
    When an agent calls a service API and the service needs a human decision, the service returns
    HTTP 202 with a `hitl` object. The agent forwards the review URL to the human. The human
    decides in a browser. The agent polls for the result.

    **Framework-agnostic.** This spec works with any HTTP server (Express, Hono, Next.js, FastAPI,
    Spring Boot, Go net/http, or any other).

    For the full specification, see: https://github.com/rotorstar/hitl-protocol
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: HITL Protocol
    url: https://github.com/rotorstar/hitl-protocol
servers:
  - url: https://api.example.com/v1
    description: Example service API
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        Agent-to-service authentication. The agent authenticates with a bearer token
        obtained through the service's standard auth flow.
    ReviewToken:
      type: apiKey
      in: query
      name: token
      description: |
        Opaque bearer token in the review URL (43 chars, base64url-encoded).
        Generated by the service: `randomBytes(32).toString('base64url')`.
        Verified by the service: `timingSafeEqual(SHA256(token), stored_hash)`.

  # ============================================================
  # Headers
  # ============================================================
  headers:
    Retry-After:
      description: Seconds until the agent should poll again. RECOMMENDED 30-300.
      schema:
        type: integer
        minimum: 1
        examples: [30]
    ETag:
      description: Opaque identifier for the current poll response version. Enables efficient polling with If-None-Match.
      schema:
        type: string
        examples: ['"v3-completed"']
    X-RateLimit-Limit:
      description: Maximum requests per minute.
      schema:
        type: integer
        examples: [60]
    X-RateLimit-Remaining:
      description: Remaining requests in the current window.
      schema:
        type: integer
        examples: [58]

  # ============================================================
  # Parameters
  # ============================================================
  parameters:
    caseId:
      name: caseId
      in: path
      required: true
      description: Unique, URL-safe review case identifier.
      schema:
        type: string
        pattern: '^[a-zA-Z0-9_-]+$'
        examples: ['review_abc123']
    reviewToken:
      name: token
      in: query
      required: true
      description: Opaque bearer token from the review URL.
      schema:
        type: string
        minLength: 43
        maxLength: 43
        examples: ['K7xR2mN4pQ8sT1vW3xY5zA9bC6dE8fG0hI2jK4lM6nO']
...
    ifNoneMatch:
      name: If-None-Match
      in: header
      required: false
      description: ETag from a previous poll response. Returns 304 if unchanged.
      schema:
        type: string
        examples: ['"v2-in_progress"']

  # ============================================================
  # Schemas (referencing existing JSON Schema files)
  # ============================================================
  schemas:
...
    PollResponse:
      $ref: './poll-response.schema.json'
...
    ReviewSubmission:
      type: object
      required:
        - action
      properties:
        action:
          type: string
          description: |
            The action taken by the human. Valid values depend on the review type:
            - **approval**: approve, edit, reject
            - **selection**: select
            - **input**: submit
            - **confirmation**: confirm, cancel
            - **escalation**: retry, skip, abort
          examples: ['approve', 'select', 'submit', 'confirm', 'retry']
        data:
          type: object
          additionalProperties: true
          description: Structured data accompanying the action.
      description: Human's response submitted via the review page.
    ReviewSubmissionResult:
      type: object
      required:
        - status
        - case_id
      properties:
        status:
          type: string
          const: completed
        case_id:
          type: string
        completed_at:
          type: string
          format: date-time
        next_case_id:
          type: string
          description: If the action triggers a follow-up (e.g., edit → new draft), the next case ID.
    DiscoveryResponse:
      type: object
      required:
        - hitl_protocol
      properties:
        hitl_protocol:
          type: object
          required:
            - spec_version
          properties:
            spec_version:
              type: string
              examples: ['0.5']
            service:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                url:
                  type: string
                  format: uri
                contact:
                  type: string
            capabilities:
              type: object
              properties:
                review_types:
                  type: array
                  items:
                    type: string
                transports:
                  type: array
                  items:
                    type: string
                    enum: [polling, sse, callback]
                max_timeout:
                  type: string
                default_timeout:
                  type: string
                supports_reminders:
                  type: boolean
                supports_multi_round:
                  type: boolean
                supports_signatures:
                  type: boolean
            endpoints:
              type: object
              properties:
                reviews_base:
                  type: string
                  format: uri
                review_page_base:
                  type: string
                  format: uri
                events_base:
                  type: string
                  format: uri
            rate_limits:
              type: object
              properties:
                poll_recommended_interval_seconds:
                  type: integer
                max_requests_per_minute:
                  type: integer
            policies:
              type: object
              properties:
                data_retention_days:
                  type: integer
                privacy_policy:
                  type: string
                  format: uri
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Machine-readable error code.
          examples: ['duplicate_submission', 'rate_limited', 'invalid_token', 'case_expired']
        message:
          type: string
          description: Human-readable error message.

  # ============================================================
  # Response Objects
  # ============================================================
  responses:
    NotModified:
      description: Poll response unchanged since last ETag. Agent should wait and retry.
    Conflict:
      description: |
        Duplicate submission. The human has already responded to this review case.
        Review cases accept exactly one response (one-time response guarantee).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: duplicate_submission
            message: This review case has already been responded to.
    RateLimited:
      description: |
        Too many requests. RECOMMENDED limit: 60 requests per minute per case.
      headers:
        Retry-After:
          $ref: '#/components/headers/Retry-After'
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/X-RateLimit-Remaining'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: rate_limited
            message: Too many requests. Please wait 30 seconds.
    InvalidToken:
      description: The review token is missing, malformed, or does not match any case.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: invalid_token
            message: Invalid or expired review token.
    CaseExpired:
      description: The review case has expired and can no longer accept responses.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: case_expired
            message: This review case expired on 2026-02-23T10:00:00Z.

# ============================================================
# Paths
# ============================================================
paths:
...
  /reviews/{caseId}/status:
    get:
      tags: [Polling]
      summary: Poll review case status
      description: |
        Agent polls this endpoint to check the current status of a review case.
        Returns one of 6 statuses: `pending`, `opened`, `in_progress`, `completed`, `expired`, `cancelled`.

        **Efficient polling:** Use `If-None-Match` with the `ETag` from a previous response.
        If the status hasn't changed, the service returns 304 Not Modified with no body.

        **Polling interval:** RECOMMENDED 30 seconds to 5 minutes. Respect `Retry-After` header.
      operationId: pollReviewStatus
      parameters:
        - $ref: '#/components/parameters/caseId'
        - $ref: '#/components/parameters/ifNoneMatch'
      responses:
        '200':
          description: Current review case status.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollResponse'
              examples:
                pending:
                  summary: Case created, human hasn't opened yet
                  value:
                    status: pending
                    case_id: review_abc123
                    created_at: "2026-02-22T10:00:00Z"
                    expires_at: "2026-02-23T10:00:00Z"
                opened:
                  summary: Human opened the review URL
                  value:
                    status: opened
                    case_id: review_abc123
                    created_at: "2026-02-22T10:00:00Z"
                    opened_at: "2026-02-22T11:15:00Z"
                    expires_at: "2026-02-23T10:00:00Z"
                in_progress:
                  summary: Human is interacting (multi-step form progress)
                  value:
                    status: in_progress
                    case_id: review_abc123
                    created_at: "2026-02-22T10:00:00Z"
                    opened_at: "2026-02-22T11:15:00Z"
                    expires_at: "2026-02-23T10:00:00Z"
                    progress:
                      current_step: 2
                      total_steps: 3
                      completed_fields: 4
                      total_fields: 9
                completed:
                  summary: Human submitted their response
                  value:
                    status: completed
                    case_id: review_abc123
                    created_at: "2026-02-22T10:00:00Z"
                    opened_at: "2026-02-22T11:15:00Z"
                    completed_at: "2026-02-22T11:19:46Z"
                    expires_at: "2026-02-23T10:00:00Z"
                    result:
                      action: select
                      data:
                        selected: [job_001, job_003]
                    responded_by:
                      name: Alex Mueller
                      email: alex@example.com
                expired:
                  summary: Case expired without response
                  value:
                    status: expired
                    case_id: review_abc123
                    created_at: "2026-02-22T10:00:00Z"
                    expires_at: "2026-02-23T10:00:00Z"
                    expired_at: "2026-02-23T10:00:00Z"
                    default_action: skip
                cancelled:
                  summary: Human cancelled the review
                  value:
                    status: cancelled
                    case_id: review_abc123
                    created_at: "2026-02-22T10:00:00Z"
                    opened_at: "2026-02-22T11:15:00Z"
                    cancelled_at: "2026-02-22T11:16:30Z"
                    reason: "Changed my mind about the application."
        '304':
          $ref: '#/components/responses/NotModified'
        '429':
          $ref: '#/components/responses/RateLimited'
  /reviews/{caseId}/respond:
    post:
      tags: [Human Review]
      summary: Submit human response
      description: |
        The review page submits the human's decision to this endpoint.
        Authenticated via the opaque bearer token from the review URL.

        **One-time response:** Each case accepts exactly one response. Subsequent submissions
        return HTTP 409 Conflict. A RECOMMENDED 5-minute grace period allows edits.
      operationId: submitReviewResponse
      security:
        - ReviewToken: []
      parameters:
        - $ref: '#/components/parameters/caseId'
        - $ref: '#/components/parameters/reviewToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewSubmission'
            examples:
              approve:
                summary: Approve a deployment
                value:
                  action: approve
                  data:
                    feedback: "LGTM, deploy to production."
              select:
                summary: Select jobs to apply for
                value:
                  action: select
                  data:
                    selected: [job_001, job_003]
                    note: "Prioritize the senior role."
              submit:
                summary: Submit form data
                value:
                  action: submit
                  data:
                    salary_expectation: 108000
                    earliest_start_date: "2026-05-01"
                    work_authorization: blue_card
              confirm:
                summary: Confirm sending emails
                value:
                  action: confirm
                  data:
                    confirmed_items: [email_1, email_2, email_3]
              retry:
                summary: Retry with modified params
                value:
                  action: retry
                  data:
                    reason: "Increase memory allocation"
                    modified_params:
                      memory: "4GB"
      responses:
        '200':
          description: Response accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewSubmissionResult'
              example:
                status: completed
                case_id: review_abc123
                completed_at: "2026-02-22T11:19:46Z"
        '401':
          $ref: '#/components/responses/InvalidToken'
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          $ref: '#/components/responses/CaseExpired'
  /reviews/{caseId}/events:
    get:
      tags: [Events]
      summary: SSE event stream (optional)
      description: |
        Optional Server-Sent Events (SSE) endpoint for real-time status updates.
        Services that offer SSE MUST also support polling.

        **Events:**
        - `review.opened` — Human opened the review URL
        - `review.in_progress` — Human started interacting
        - `review.completed` — Human submitted response (includes result)
        - `review.expired` — Case expired
        - `review.cancelled` — Human cancelled
        - `review.reminder` — Reminder sent

        **Reconnection:** Clients should use `Last-Event-ID` header on reconnect.
        If SSE connection fails, fall back to polling.
      operationId: reviewEventStream
      parameters:
        - $ref: '#/components/parameters/caseId'
        - name: Last-Event-ID
          in: header
          required: false
          schema:
            type: string
          description: Resume stream from this event ID after reconnection.
      responses:
        '200':
          description: SSE event stream.
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  Server-Sent Events stream. Each event has:
                  - `event:` field (event type)
                  - `data:` field (JSON payload)
                  - `id:` field (event ID for reconnection)
              example: |
                event: review.opened
                data: {"case_id":"review_abc123","opened_at":"2026-02-22T11:15:00Z"}
                id: evt_001

                event: review.completed
                data: {"case_id":"review_abc123","status":"completed","result":{"action":"select","data":{"selected":["job_001"]}}}
                id: evt_002
        '401':
          description: Agent authentication failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /.well-known/hitl.json:
    get:
      tags: [Discovery]
      summary: HITL Protocol discovery endpoint
      description: |
        Returns the service's HITL Protocol capabilities, supported review types,
        transport mechanisms, rate limits, and endpoint locations.

        Agents can use this endpoint to discover what a service supports before
        making API requests. RECOMMENDED cache with `Cache-Control: public, max-age=86400`.
      operationId: getDiscovery
      security: []
      responses:
        '200':
          description: Service HITL capabilities.
        '404':
          description: Service does not support HITL Protocol discovery.
          headers:
            Cache-Control:
              schema:
                type: string
                examples: ['public, max-age=86400']
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscoveryResponse'
              example:
                hitl_protocol:
                  spec_version: "0.5"
                  service:
                    name: Example Service
                    description: HITL-enabled service
                    url: https://service.example
                  capabilities:
                    review_types: [approval, selection, input, confirmation, escalation]
                    transports: [polling, sse]
                    default_timeout: PT24H
                    supports_reminders: true
                    supports_multi_round: true
                  endpoints:
                    reviews_base: https://api.service.example/v1/reviews
                    review_page_base: https://service.example/review
                  rate_limits:
                    poll_recommended_interval_seconds: 30
                    max_requests_per_minute: 60

# ============================================================
# Webhooks (Callback transport — OpenAPI 3.1 feature)
# ============================================================
...